switch cli entirely to plugins
