+++ src/cli_plugins/plugins.json
@@ -7,6 +7,56 @@
                 "chart",
                 "flowchart"
             ]
+        },
+        {
+            "module": "code",
+            "function": "code",
+            "subcommands": [
+                "code",
+                "vscode"
+            ]
+        },
+        {
+            "module": "data",
+            "function": "data",
+            "subcommands": [
+                "data"
+            ]
+        },
+        {
+            "module": "format_files",
+            "function": "format_files",
+            "subcommands": [
+                "format"
+            ]
+        },
+        {
+            "module": "install",
+            "function": "install",
+            "subcommands": [
+                "install"
+            ]
+        },
+        {
+            "module": "lint",
+            "function": "lint",
+            "subcommands": [
+                "lint"
+            ]
+        },
+        {
+            "module": "run",
+            "function": "run",
+            "subcommands": [
+                "run"
+            ]
+        },
+        {
+            "module": "shell",
+            "function": "shell",
+            "subcommands": [
+                "shell"
+            ]
         }
     ]
 }
