+++ src/cli_plugins/run.py
@@ -1,0 +1,8 @@
+def run():
+    """
+    Starts the baboon tracker algorithm.
+    """
+
+    from baboon_tracking import BaboonTracker  # pylint: disable=import-outside-toplevel
+
+    BaboonTracker().run()
